Move the MergeFunction trait into its own module
